correction de type de annonce_id

+++ annonce-api/src/main/java/fr/checkconsulting/annonceapi/entity/Annonce.java
@@ -17,7 +17,7 @@ import java.time.LocalDate;
 public class Annonce {
 
     @Id
-    private Long annonceId;
+    private Integer annonceId;
     private String titre;
     private String description;
     private String categorie;
